+++ st2common/tests/integration/test_logging.py
@@ -18,6 +18,7 @@ from __future__ import absolute_import
 import os
 import sys
 import signal
+import unittest
 
 import eventlet
 from eventlet.green import subprocess
@@ -30,7 +31,9 @@ TEST_FILE_PATH = os.path.join(BASE_DIR, "log_unicode_data.py")
 
 
 class LogFormattingAndEncodingTestCase(IntegrationTestCase):
-    def test_formatting_with_unicode_data_works_no_stdout_patching(self):
+    def test_formatting_with_unicode_data_works_no_stdout_patching_valid_utf8_encoding(
+        self,
+    ):
         # Ensure that process doesn't end up in an infinite loop if non-utf8 locale / encoding is
         # used and a unicode sequence is logged.
 
@@ -65,6 +68,13 @@ class LogFormattingAndEncodingTestCase(IntegrationTestCase):
         self.assertIn(
             "DEBUG [-] Test debug message with unicode 1 - \u597d\u597d\u597d", stdout
         )
+
+    @unittest.skipIf(sys.version_info >= (3, 8, 0), "Skipping test under Python >= 3.8")
+    def test_formatting_with_unicode_data_works_no_stdout_patching_non_valid_utf8_encoding(
+        self,
+    ):
+        # Ensure that process doesn't end up in an infinite loop if non-utf8 locale / encoding is
+        # used and a unicode sequence is logged.
 
         # 2. Process is not using utf-8 encoding - LC_ALL set to invalid locale - should result in
         # single exception being logged, but not infinite loop
@@ -105,6 +115,12 @@ class LogFormattingAndEncodingTestCase(IntegrationTestCase):
             "DEBUG [-] Test debug message with unicode 1 - \u597d\u597d\u597d", stdout
         )
 
+    def test_formatting_with_unicode_data_works_no_stdout_patching_ascii_pythonioencoding(
+        self,
+    ):
+        # Ensure that process doesn't end up in an infinite loop if non-utf8 locale / encoding is
+        # used and a unicode sequence is logged.
+
         # 3. Process is not using utf-8 encoding - PYTHONIOENCODING set to ascii - should result in
         # single exception being logged, but not infinite loop
         process = self._start_process(
@@ -144,7 +160,9 @@ class LogFormattingAndEncodingTestCase(IntegrationTestCase):
             "DEBUG [-] Test debug message with unicode 1 - \u597d\u597d\u597d", stdout
         )
 
-    def test_formatting_with_unicode_data_works_with_stdout_patching(self):
+    def test_formatting_with_unicode_data_works_with_stdout_patching_valid_utf8_encoding(
+        self,
+    ):
         # Test a scenario where patching is enabled which means it should never result in infinite
         # loop
         # 1. Process is using a utf-8 encoding
@@ -164,7 +182,6 @@ class LogFormattingAndEncodingTestCase(IntegrationTestCase):
         stdout = process.stdout.read().decode("utf-8")
         stderr = process.stderr.read().decode("utf-8")
         stdout_lines = stdout.split("\n")
-        print(stderr)
 
         self.assertEqual(stderr, "")
         self.assertTrue(len(stdout_lines) < 20)
@@ -180,6 +197,9 @@ class LogFormattingAndEncodingTestCase(IntegrationTestCase):
             "DEBUG [-] Test debug message with unicode 1 - \u597d\u597d\u597d", stdout
         )
 
+    def test_formatting_with_unicode_data_works_with_stdout_patching_non_valid_utf8_encoding(
+        self,
+    ):
         # 2. Process is not using utf-8 encoding
         process = self._start_process(
             env={
@@ -196,9 +216,11 @@ class LogFormattingAndEncodingTestCase(IntegrationTestCase):
 
         stdout = process.stdout.read().decode("utf-8")
         stderr = process.stderr.read().decode("utf-8")
+        stdout_lines = stdout.split("\n")
 
         self.assertEqual(stderr, "")
-        self.assertTrue(len(stdout_lines) < 50)
+        print(stdout)
+        self.assertTrue(len(stdout_lines) < 100)
 
         self.assertIn("INFO [-] Test info message 1", stdout)
         self.assertIn("Test debug message 1", stdout)
@@ -211,6 +233,9 @@ class LogFormattingAndEncodingTestCase(IntegrationTestCase):
             "DEBUG [-] Test debug message with unicode 1 - \u597d\u597d\u597d", stdout
         )
 
+    def test_formatting_with_unicode_data_works_with_stdout_patching__ascii_pythonioencoding(
+        self,
+    ):
         # 3. Process is not using utf-8 encoding - PYTHONIOENCODING set to ascii
         process = self._start_process(
             env={