+++ st2common/tests/unit/test_db.py
@@ -103,10 +103,12 @@ class DbConnectionTestCase(DbTestCase):
     def setUp(self):
         # NOTE: It's important we re-establish a connection on each setUp
         self.setUpClass()
+        cfg.CONF.reset()
 
     def tearDown(self):
         # NOTE: It's important we disconnect here otherwise tests will fail
         disconnect()
+        cfg.CONF.reset()
 
     def test_check_connect(self):
         """
@@ -120,6 +122,108 @@ class DbConnectionTestCase(DbTestCase):
             cfg.CONF.database.port,
         )
         self.assertIn(expected_str, str(client), "Not connected to desired host.")
+
+    def test_network_level_compression(self):
+        disconnect()
+
+        db_name = "st2"
+        db_host = "localhost"
+        db_port = 27017
+
+        # If running version < MongoDB 4.2 we skip this check since zstd is only supported in server
+        # >= 4.2
+        connection = db_setup(
+            db_name=db_name,
+            db_host=db_host,
+            db_port=db_port,
+            ensure_indexes=False,
+        )
+        server_version = tuple(
+            [int(x) for x in connection.server_info()["version"].split(".")]
+        )
+
+        if server_version < (4, 2, 0):
+            self.skipTest("Skipping test since running MongoDB < 4.2")
+            return
+
+        disconnect()
+
+        # 1. Verify default is no compression
+        connection = db_setup(
+            db_name=db_name,
+            db_host=db_host,
+            db_port=db_port,
+            ensure_indexes=False,
+        )
+        # Sadly there is no nicer way to assert that it seems
+        self.assertFalse("compressors=['zstd']" in str(connection))
+        self.assertFalse("compressors" in str(connection))
+
+        # 2. Verify using zstd works - specified using config option
+        disconnect()
+
+        cfg.CONF.set_override(name="compressors", group="database", override="zstd")
+
+        connection = db_setup(
+            db_name=db_name,
+            db_host=db_host,
+            db_port=db_port,
+            ensure_indexes=False,
+        )
+        # Sadly there is no nicer way to assert that it seems
+        self.assertTrue("compressors=['zstd']" in str(connection))
+
+        # 3. Verify using zstd works - specified inside URI
+        disconnect()
+
+        cfg.CONF.set_override(name="compressors", group="database", override=None)
+        db_host = "mongodb://127.0.0.1/?compressors=zstd"
+
+        connection = db_setup(
+            db_name=db_name,
+            db_host=db_host,
+            db_port=db_port,
+            ensure_indexes=False,
+        )
+        # Sadly there is no nicer way to assert that it seems
+        self.assertTrue("compressors=['zstd']" in str(connection))
+
+        # 4. Verify using zlib works - specified using config option
+        disconnect()
+
+        cfg.CONF.set_override(name="compressors", group="database", override="zlib")
+        cfg.CONF.set_override(
+            name="zlib_compression_level", group="database", override=8
+        )
+
+        connection = db_setup(
+            db_name=db_name,
+            db_host=db_host,
+            db_port=db_port,
+            ensure_indexes=False,
+        )
+        # Sadly there is no nicer way to assert that it seems
+        self.assertTrue("compressors=['zlib']" in str(connection))
+        self.assertTrue("zlibcompressionlevel=8" in str(connection))
+
+        # 5. Verify using zlib works - specified inside URI
+        disconnect()
+
+        cfg.CONF.set_override(name="compressors", group="database", override=None)
+        cfg.CONF.set_override(
+            name="zlib_compression_level", group="database", override=None
+        )
+        db_host = "mongodb://127.0.0.1/?compressors=zlib&zlibCompressionLevel=9"
+
+        connection = db_setup(
+            db_name=db_name,
+            db_host=db_host,
+            db_port=db_port,
+            ensure_indexes=False,
+        )
+        # Sadly there is no nicer way to assert that it seems
+        self.assertTrue("compressors=['zlib']" in str(connection))
+        self.assertTrue("zlibcompressionlevel=9" in str(connection))
 
     def test_get_ssl_kwargs(self):
         # 1. No SSL kwargs provided
@@ -209,6 +313,7 @@ class DbConnectionTestCase(DbTestCase):
             username="username",
             password="password",
             authentication_mechanism="MONGODB-X509",
+            ensure_indexes=False,
         )
 
         call_args = mock_mongoengine.connection.connect.call_args_list[0][0]
@@ -250,6 +355,7 @@ class DbConnectionTestCase(DbTestCase):
             db_port=db_port,
             username=username,
             password=password,
+            ensure_indexes=False,
         )
 
         expected_message = (
@@ -276,6 +382,7 @@ class DbConnectionTestCase(DbTestCase):
             db_port=db_port,
             username=username,
             password=password,
+            ensure_indexes=False,
         )
 
         expected_message = (
@@ -302,6 +409,7 @@ class DbConnectionTestCase(DbTestCase):
             db_port=db_port,
             username=username,
             password=password,
+            ensure_indexes=False,
         )
 
         expected_message = (
@@ -327,6 +435,7 @@ class DbConnectionTestCase(DbTestCase):
             db_port=db_port,
             username=username,
             password=password,
+            ensure_indexes=False,
         )
 
         expected_message = (
@@ -352,6 +461,7 @@ class DbConnectionTestCase(DbTestCase):
             db_port=db_port,
             username=username,
             password=password,
+            ensure_indexes=False,
         )
 
         expected_message = (
@@ -390,6 +500,7 @@ class DbConnectionTestCase(DbTestCase):
             db_port=db_port,
             username=username,
             password=password,
+            ensure_indexes=False,
         )
 
         expected_message = (
@@ -414,9 +525,7 @@ class DbConnectionTestCase(DbTestCase):
         db_host = "localhost"
         db_port = 27017
 
-        cfg.CONF.set_override(
-            name="connection_timeout", group="database", override=1000
-        )
+        cfg.CONF.set_override(name="connection_timeout", group="database", override=300)
 
         start = time.time()
         self.assertRaises(
@@ -426,15 +535,16 @@ class DbConnectionTestCase(DbTestCase):
             db_host=db_host,
             db_port=db_port,
             ssl=True,
+            ensure_indexes=False,
         )
         end = time.time()
         diff = end - start
 
-        self.assertTrue(diff >= 1)
+        self.assertTrue(diff >= 0.3)
 
         disconnect()
 
-        cfg.CONF.set_override(name="connection_timeout", group="database", override=400)
+        cfg.CONF.set_override(name="connection_timeout", group="database", override=200)
 
         start = time.time()
         self.assertRaises(
@@ -444,11 +554,12 @@ class DbConnectionTestCase(DbTestCase):
             db_host=db_host,
             db_port=db_port,
             ssl=True,
+            ensure_indexes=False,
         )
         end = time.time()
         diff = end - start
 
-        self.assertTrue(diff >= 0.4)
+        self.assertTrue(diff >= 0.1)
 
 
 class DbCleanupTestCase(DbTestCase):