+++ CHANGELOG.rst
@@ -66,7 +66,7 @@ Added
   Contributed by @Kami.
 
 * Add support for compressing the payloads which are sent over the message bus. Compression is
-  disabled by default and user can enable it by setting ``transport.compression`` config option
+  disabled by default and user can enable it by setting ``messaging.compression`` config option
   to one of the following values: ``zstd``, ``lzma``, ``bz2``, ``gzip``.
 
   In most cases we recommend using ``zstd`` (zstandard) since it offers best trade off between